other: rewrote how plot pages work and call their data, so that the allele selection options persist between genomic and translated plots.
